consolidate merge utilities
